+++ apps/explorer/lib/explorer/chain.ex
@@ -1823,6 +1823,15 @@ defmodule Explorer.Chain do
   end
 
   @doc """
+  Finds all transactions of a certain block numbers
+  """
+  def get_transactions_of_block_numbers(block_numbers) do
+    block_numbers
+    |> Transaction.transactions_for_block_numbers()
+    |> Repo.all()
+  end
+
+  @doc """
   Finds all Blocks validated by the address with the given hash.
 
     ## Options
@@ -1946,86 +1955,6 @@ defmodule Explorer.Chain do
   def stream_unfetched_token_balances(initial, reducer, limited? \\ false) when is_function(reducer, 2) do
     TokenBalance.unfetched_token_balances()
     |> add_token_balances_fetcher_limit(limited?)
-    |> Repo.stream_reduce(initial, reducer)
-  end
-
-  @doc """
-  Returns a stream of all blocks with unfetched internal transactions, using
-  the `pending_block_operation` table.
-
-      iex> unfetched = insert(:block)
-      iex> insert(:pending_block_operation, block: unfetched, block_number: unfetched.number)
-      iex> {:ok, number_set} = Explorer.Chain.stream_blocks_with_unfetched_internal_transactions(
-      ...>   MapSet.new(),
-      ...>   fn number, acc ->
-      ...>     MapSet.put(acc, number)
-      ...>   end
-      ...> )
-      iex> unfetched.number in number_set
-      true
-
-  """
-  @spec stream_blocks_with_unfetched_internal_transactions(
-          initial :: accumulator,
-          reducer :: (entry :: term(), accumulator -> accumulator),
-          limited? :: boolean()
-        ) :: {:ok, accumulator}
-        when accumulator: term()
-  def stream_blocks_with_unfetched_internal_transactions(initial, reducer, limited? \\ false)
-      when is_function(reducer, 2) do
-    direction = Application.get_env(:indexer, :internal_transactions_fetch_order)
-
-    query =
-      from(
-        po in PendingBlockOperation,
-        where: not is_nil(po.block_number),
-        select: po.block_number,
-        order_by: [{^direction, po.block_number}]
-      )
-
-    query
-    |> add_fetcher_limit(limited?)
-    |> Repo.stream_reduce(initial, reducer)
-  end
-
-  @doc """
-  Returns a stream of all transactions with unfetched internal transactions, using
-  the `pending_transaction_operation` table.
-      iex> unfetched_block = insert(:block)
-      iex> unfetched_transaction = insert(:transaction) |> with_block(unfetched_block)
-      iex> insert(:pending_transaction_operation, transaction: unfetched_transaction)
-      iex> {:ok, transaction_params_set} = Explorer.Chain.stream_transactions_with_unfetched_internal_transactions(
-      ...>   MapSet.new(),
-      ...>   fn transaction_params, acc ->
-      ...>     MapSet.put(acc, transaction_params)
-      ...>   end
-      ...> )
-      iex> %{
-      ...>   block_number: unfetched_transaction.block_number,
-      ...>   hash: unfetched_transaction.hash,
-      ...>   index: unfetched_transaction.index
-      ...> } in transaction_params_set
-      true
-  """
-  @spec stream_transactions_with_unfetched_internal_transactions(
-          initial :: accumulator,
-          reducer :: (entry :: term(), accumulator -> accumulator)
-        ) :: {:ok, accumulator}
-        when accumulator: term()
-  def stream_transactions_with_unfetched_internal_transactions(initial, reducer, limited? \\ false)
-      when is_function(reducer, 2) do
-    direction = Application.get_env(:indexer, :internal_transactions_fetch_order)
-
-    query =
-      from(
-        po in PendingTransactionOperation,
-        join: t in assoc(po, :transaction),
-        select: %{block_number: t.block_number, hash: t.hash, index: t.index},
-        order_by: [{^direction, t.block_number}]
-      )
-
-    query
-    |> add_fetcher_limit(limited?)
     |> Repo.stream_reduce(initial, reducer)
   end
